main: Fix cloud-init and NetworkManager detection
Neither `which` or `type` do a good job at detecting those binaries.
Also, on some distros, $PATH does not include all relevant paths.
Therefore, we should perform the checks manually.

+++ distrobuilder/main.go
@@ -506,6 +506,16 @@ is_lxd_vm() {
 	[ -e /dev/virtio-ports/org.linuxcontainers.lxd ]
 }
 
+# is_in_path succeeds if the given file exists in on of the paths
+is_in_path() {
+	# Don't use $PATH as that may not include all relevant paths
+	for path in /bin /sbin /usr/bin /usr/sbin /usr/local/bin /usr/local/sbin; do
+		[ -e "${path}/$1" ] && return 0
+	done
+
+	return 1
+}
+
 ## Fix functions
 # fix_ro_paths avoids udevd issues with /sys and /proc being writable
 fix_ro_paths() {
@@ -522,7 +532,7 @@ fix_nm_force_up() {
 	[ -e "/sys/class/net/$1" ] || return 0
 
 	# Check if NetworkManager exists
-	type NetworkManager >/dev/null || return 0
+	[ "${nm_exists}" -eq 1 ] || return 0
 
 	cat <<-EOF > /run/systemd/system/network-connection-activate.service
 [Unit]
@@ -596,6 +606,13 @@ if ! is_lxd_vm && ! is_lxc_container; then
 	exit
 fi
 
+# Check for NetworkManager and cloud-init
+nm_exists=0
+cloudinit_exists=0
+
+is_in_path NetworkManager && nm_exists=1
+is_in_path cloud-init && cloudinit_exists=1
+
 # Determine systemd version
 for path in /usr/lib/systemd/systemd /lib/systemd/systemd; do
 	[ -x "${path}" ] || continue
@@ -631,12 +648,12 @@ if is_lxc_container; then
 fi
 
 # Workarounds for fedora/34/cloud containers
-if is_lxc_container && [ "${ID}" = "fedora" ] && [ "${VERSION_ID}" = "34" ] && type cloud-init >/dev/null; then
+if is_lxc_container && [ "${ID}" = "fedora" ] && [ "${VERSION_ID}" = "34" ] && [ "${cloudinit_exists}" -eq 1 ]; then
 	fix_nm_force_up eth0
 fi
 
 # Workarounds for NetworkManager in containers
-if type NetworkManager >/dev/null; then
+if [ "${nm_exists}" -eq 1 ]; then
 	fix_nm_link_state eth0
 fi
 `
